Change code so that it accepts double value instead of int.

+++ basicCS/FactorialSolver/FactorialSolver/FactorialSolver.cpp
@@ -8,12 +8,12 @@
 
 using namespace std; 
 
-long int factorial(long int n); 
+double factorial(double n); 
 
 int main(){
 	cout << " ------------------ Factorial Solver ---------------------" << endl; 	
 
-	long int userInput = 0; // Holds user input 
+	double userInput = 0; // Holds user input 
 
 	cout << "Enter a number to find its factorial (enter -1 to exit): "; 
 	cin >> userInput; 
@@ -29,7 +29,7 @@ int main(){
 
 }
 
-long int factorial(long int n)
+double factorial(double n)
 {
 	if( n == 0 || n == 1) // Base case
 		return 1;
